project_b/edgeimpulse/requirements.txt
tensorflow-model-optimization==0.8.0
